+++ rally_openstack/contexts/cleanup/admin.py
@@ -36,6 +36,5 @@ class AdminCleanup(base.CleanupMixin, context.Context):
             admin_required=True,
             admin=self.context["admin"],
             users=self.context.get("users", []),
-            api_versions=self.context["config"].get("api_versions"),
             superclass=scenario.OpenStackScenario,
             task_id=self.get_owner_id())