+++ rally_openstack/contexts/quotas/quotas.py
@@ -50,8 +50,7 @@ class Quotas(context.Context):
     def __init__(self, ctx):
         super(Quotas, self).__init__(ctx)
         self.clients = osclients.Clients(
-            self.context["admin"]["credential"],
-            api_info=self.context["config"].get("api_versions"))
+            self.context["admin"]["credential"])
 
         self.manager = {
             "nova": nova_quotas.NovaQuotas(self.clients),